requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
# psycopg
# psycopg2-binary
asyncpg-lite
psycopg2==2.9.9
geopy
requests
pytz
timezonefinder